taihe: implement file validation
Issue: https://gitee.com/openharmony/arkui_ace_engine/issues/ICTXCB

+++ taihe/process_sdk_files.py
@@ -16,7 +16,6 @@
 
 import os
 import shutil
-import stat
 import argparse
 from pathlib import Path
 
@@ -31,9 +30,10 @@ def should_skip_file(file_path, patterns) -> bool:
 
 
 def copy_with_permissions(src, dst):
-    shutil.copy2(src, dst)
-    src_stat = os.stat(src)
-    os.chmod(dst, src_stat.st_mode)
+    if src.exists():
+        shutil.copy2(src, dst)
+        src_stat = os.stat(src)
+        os.chmod(dst, src_stat.st_mode)
 
 
 def copy_directory_excluding_files(base_dir, output_dir):
